requirements.txt
streamlit
google-generativeai
folium
geopy
requests
pyngrok
